Factor out style data
